requirements.txt
Flask==3.1.0
google-generativeai==0.8.3
gunicorn==23.0.0
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.1
qdrant-client==1.12.1
--find-links https://download.pytorch.org/whl/torch/
torch==2.5.1+cpu
